# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TICKET API",
        "description": "Documentation de mon API Laravel",
        "contact": {
            "name": "Support API",
            "email": "shadracmeshacdimon@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Liste paginée des événements en cours",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "responses": {
                    "200": {
                        "description": "Liste des événements en cours récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Event"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string"
                                        },
                                        "from": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "last_page_url": {
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        },
                                        "to": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Créer un nouvel événement",
                "operationId": "508895e03d7f10edac84357084c87b20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "event_title",
                                    "event_description",
                                    "event_date",
                                    "event_image",
                                    "event_category",
                                    "event_city",
                                    "event_address",
                                    "event_status"
                                ],
                                "properties": {
                                    "event_title": {
                                        "type": "string",
                                        "maxLength": 30
                                    },
                                    "event_description": {
                                        "type": "string",
                                        "maxLength": 16777215
                                    },
                                    "event_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "event_image": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "event_category": {
                                        "type": "string",
                                        "enum": [
                                            "Autre",
                                            "Concert-Spectacle",
                                            "Diner Gala",
                                            "Festival",
                                            "Formation"
                                        ]
                                    },
                                    "event_city": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "event_address": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "event_status": {
                                        "type": "string",
                                        "enum": [
                                            "upcoming",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Événement créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Afficher un événement spécifique",
                "operationId": "6f743f71f7a88dbe57be1d9c9ebf47cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Événement récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Event"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Événement non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Event": {
                "required": [
                    "event_id",
                    "event_title",
                    "event_description",
                    "event_date",
                    "event_image",
                    "event_category",
                    "event_city",
                    "event_address",
                    "event_status"
                ],
                "properties": {
                    "event_id": {
                        "type": "integer"
                    },
                    "event_title": {
                        "type": "string",
                        "maxLength": 30
                    },
                    "event_description": {
                        "type": "string",
                        "maxLength": 16777215
                    },
                    "event_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "event_image": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "event_category": {
                        "type": "string",
                        "enum": [
                            "Autre",
                            "Concert-Spectacle",
                            "Diner Gala",
                            "Festival",
                            "Formation"
                        ]
                    },
                    "event_city": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "event_address": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "event_status": {
                        "type": "string",
                        "enum": [
                            "upcoming",
                            "completed",
                            "cancelled"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Use the token provided after logging in.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}